# apps/api/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthenticationErrorDto": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "AUTHENTICATION_REQUIRED",
              "AUTHENTICATION_INVALID"
            ],
            "type": "string"
          },
          "message": {
            "description": "Safe, localizable public message.",
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "CreateHouseholdRequestDto": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "description": "Human-readable Household name.",
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "HouseholdDetailDto": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "description": "Opaque Household identifier.",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "membershipStatus": {
            "const": "active",
            "type": "string"
          },
          "name": {
            "description": "Human-readable Household name.",
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "member"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "membershipStatus",
          "name",
          "role"
        ],
        "type": "object"
      },
      "HouseholdErrorDto": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "HOUSEHOLD_FORBIDDEN",
              "HOUSEHOLD_INVITATION_NOT_FOUND",
              "HOUSEHOLD_INVITATION_UNAVAILABLE",
              "HOUSEHOLD_NOT_FOUND",
              "HOUSEHOLD_VALIDATION_FAILED"
            ],
            "type": "string"
          },
          "message": {
            "description": "Safe, localizable public message.",
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "ListHouseholdsResponseDto": {
        "additionalProperties": false,
        "properties": {
          "households": {
            "items": {
              "properties": {
                "id": {
                  "description": "Opaque Household identifier.",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "type": "string"
                },
                "membershipStatus": {
                  "const": "active",
                  "type": "string"
                },
                "name": {
                  "description": "Human-readable Household name.",
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                },
                "role": {
                  "enum": [
                    "owner",
                    "member"
                  ],
                  "type": "string"
                }
              },
              "required": [
                "id",
                "membershipStatus",
                "name",
                "role"
              ],
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "households"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "auth0": {
        "bearerFormat": "JWT",
        "description": "Auth0 access token issued for the Copiloto Financiero API audience.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "API contracts implemented so far. It contains no financial operations.",
    "title": "Copiloto Financiero API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/households": {
      "get": {
        "operationId": "HouseholdsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "households": [
                    {
                      "id": "22222222-2222-4222-8222-222222222222",
                      "membershipStatus": "active",
                      "name": "Hogar de prueba",
                      "role": "owner"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ListHouseholdsResponseDto"
                }
              }
            },
            "description": "Only active memberships for the authenticated User are returned."
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "code": "AUTHENTICATION_REQUIRED",
                  "message": "Se requiere una sesión válida."
                },
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorDto"
                }
              }
            },
            "description": "A valid access token for this API is required."
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "List active Households for the authenticated User",
        "tags": [
          "households"
        ]
      },
      "post": {
        "operationId": "HouseholdsController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHouseholdRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "id": "22222222-2222-4222-8222-222222222222",
                  "membershipStatus": "active",
                  "name": "Hogar de prueba",
                  "role": "owner"
                },
                "schema": {
                  "$ref": "#/components/schemas/HouseholdDetailDto"
                }
              }
            },
            "description": "The Household and its single initial Owner Active membership were created."
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "code": "HOUSEHOLD_VALIDATION_FAILED",
                  "message": "Revisa los datos del hogar."
                },
                "schema": {
                  "$ref": "#/components/schemas/HouseholdErrorDto"
                }
              }
            },
            "description": "The request body does not satisfy the shared Household contract."
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "code": "AUTHENTICATION_REQUIRED",
                  "message": "Se requiere una sesión válida."
                },
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorDto"
                }
              }
            },
            "description": "A valid access token for this API is required."
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Create a Household with the authenticated User as Owner Active",
        "tags": [
          "households"
        ]
      }
    },
    "/api/v1/households/{householdId}": {
      "get": {
        "operationId": "HouseholdsController_get",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "id": "22222222-2222-4222-8222-222222222222",
                  "membershipStatus": "active",
                  "name": "Hogar de prueba",
                  "role": "owner"
                },
                "schema": {
                  "$ref": "#/components/schemas/HouseholdDetailDto"
                }
              }
            },
            "description": "The authenticated User has an active membership in the Household."
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "code": "AUTHENTICATION_REQUIRED",
                  "message": "Se requiere una sesión válida."
                },
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorDto"
                }
              }
            },
            "description": "A valid access token for this API is required."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "code": "HOUSEHOLD_NOT_FOUND",
                  "message": "El hogar no existe o no está disponible."
                },
                "schema": {
                  "$ref": "#/components/schemas/HouseholdErrorDto"
                }
              }
            },
            "description": "The Household does not exist or is not visible to the authenticated User."
          }
        },
        "security": [
          {
            "auth0": []
          }
        ],
        "summary": "Get an authorized Household basic profile",
        "tags": [
          "households"
        ]
      }
    },
...
  },
  "servers": [],
...
}